Split #man out of CommandKit::Help::Man into CommandKit::Man.
