+++ lib/command_kit/help/man.rb
@@ -3,6 +3,7 @@
 require 'command_kit/command_name'
 require 'command_kit/help'
 require 'command_kit/stdio'
+require 'command_kit/man'
 
 module CommandKit
   module Help
@@ -23,6 +24,7 @@ module CommandKit
       include CommandName
       include Help
       include Stdio
+      include CommandKit::Man
 
       #
       # @api private
@@ -93,29 +95,6 @@ module CommandKit
           else
             @man_page || "#{command_name}.1"
           end
-        end
-      end
-
-      #
-      # Displays the given man page.
-      #
-      # @param [String] page
-      #   The man page file name.
-      #
-      # @param [Integer, String, nil] section
-      #   The optional section number to specify.
-      #
-      # @return [Boolean, nil]
-      #   Specifies whether the `man` command was successful or not.
-      #   Returns `nil` when the `man` command is not installed.
-      #
-      # @api public
-      #
-      def man(page, section: nil)
-        if section
-          system('man',section.to_s,page.to_s)
-        else
-          system('man',page.to_s)
         end
       end
